requirements.txt
alembic==1.16.5
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
eventlet==0.40.3
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
simple-websocket==1.1.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
wsproto==1.2.0
